Implement CameraController#takePictureToFile, CameraController#startRecording, CameraControl, CameraInfo

+++ camerax_ios/ios/Classes/video/RecordingStats.swift
@@ -9,10 +9,10 @@ import Foundation
 
 public class RecordingStats: NSObject {
     let audioStats: AudioStats
-    let numBytesRecorded: Int
-    let recorderedDurationNanos: Int
+    let numBytesRecorded: Int64
+    let recorderedDurationNanos: Int64
     
-    init(audioStats: AudioStats, numBytesRecorded: Int, recorderedDurationNanos: Int) {
+    init(audioStats: AudioStats, numBytesRecorded: Int64, recorderedDurationNanos: Int64) {
         self.audioStats = audioStats
         self.numBytesRecorded = numBytesRecorded
         self.recorderedDurationNanos = recorderedDurationNanos
